feat(locale): Port to ESM
Update the locale package to use ES module syntax 
instead of CommonJS.
