+++ packages/locale/src/config.ts
