+++ packages/locale/src/index.ts
@@ -1,0 +1,9 @@
+import { config } from "./config.js";
+
+export type Locale = keyof typeof config;
+
+export { config };
+
+export function isLocale(s: string): s is Locale {
+  return s in config;
+}
