+++ packages/locale/src/locales/de.ts
@@ -1,5 +1,5 @@
 // This file is generated by generate.js
-export default {
+export const de = {
   "accountPage": {
     "contactInformation": "Kontaktdaten",
     "deliveryAddress": "Adresse",
@@ -7,34 +7,34 @@ export default {
     "loginDetail": "Anmeldedaten",
     "mfa": {
       "codeInput": {
-        "info": "Wenn du dein 2FA-Gerät verloren hast, wende dich bitte an den Support.",
+        "info": "Wenn Sie Ihr 2FA-Gerät verloren haben, wenden Sie sich bitte an den Support.",
         "label": "Code"
       },
       "confirmDelete": {
-        "desc": "Möchtest du die Multi-Faktor-Authentifizierung wirklich deaktivieren? \nDanach brauchst du nur noch dein Passwort zum Anmelden. Bitte vergiss \nnicht, den Eintrag in deiner Authenticator-App zu löschen. Du kannst \ndiese Funktion jederzeit wieder einschalten.",
-        "descToken": "Um 2FA zu deaktivieren, gib bitte deinen aktuellen 2FA-Code ein:",
+        "desc": "Möchten Sie die Multi-Faktor-Authentifizierung wirklich deaktivieren? \nDanach benötigen Sie nur noch Ihr Passwort zur Anmeldung. Bitte \nvergessen Sie nicht, den Eintrag aus Ihrer Authenticator-App zu \nentfernen. Sie können diese Funktion jederzeit wieder aktivieren.",
+        "descToken": "Um die 2FA zu deaktivieren, geben Sie bitte Ihren aktuellen 2FA-Code ein:",
         "title": "2FA deaktivieren"
       },
-      "createUnknownErrorNotification": "Die 2-Faktor-Authentifizierung konnte aufgrund eines unbekannten Fehlers nicht gespeichert werden. Bitte versuche es erneut oder kontaktiere den Support.",
-      "deleteUnknownErrorNotification": "Die 2-Faktor-Authentifizierung konnte aufgrund eines unbekannten Fehlers nicht entfernt werden. Bitte versuche es erneut oder kontaktiere den Support.",
+      "createUnknownErrorNotification": "Die 2-Faktor-Authentifizierung konnte aufgrund eines unbekannten Fehlers\n nicht gespeichert werden. Bitte versuchen Sie es erneut oder \nkontaktieren Sie den Support.",
+      "deleteUnknownErrorNotification": "Die 2-Faktor-Authentifizierung konnte aufgrund eines unbekannten Fehlers\n nicht entfernt werden. Bitte versuchen Sie es erneut oder \nkontaktieren Sie den Support.",
       "disabledNotification": "Zwei-Faktor Authentifizierung erfolgreich deaktiviert",
       "enabledNotification": "Zwei-Faktor Authentifizierung erfolgreich aktiviert",
       "enterCode": {
-        "desc": "Gebe den Code ein, der in deiner Authentifizierungs-App angezeigt wird.",
+        "desc": "Geben Sie den Code ein, der in Ihrer Authentifizierungs-App angezeigt wird.",
         "title": "Überprüfung"
       },
       "modalTitle": "Aktivere 2-Faktor Authentifizierung",
       "result": {
-        "invalidCode": "Der Code konnte nicht bestätigt werden. Bitte versuche es erneut oder kontaktiere den Support.",
+        "invalidCode": "Der Code konnte nicht bestätigt werden. Bitte versuchen Sie es erneut oder kontaktieren Sie den Support.",
         "successful": "Die 2-Faktor Authentifizierung wurde erfolgreich eingerichtet.",
         "title": "Fertig"
       },
       "scan": {
-        "desc": "Scanne den unten stehenden QR-Code mit deiner Authentifizierungs-App.",
+        "desc": "Scannen Sie den unten stehenden QR-Code mit Ihrer Authentifizierungs-App.",
         "title": "QR-Code"
       },
       "secretInput": {
-        "desc": "Oder gib einfach den folgenden Code in deine Authenticator-App ein.",
+        "desc": "Alternativ können Sie den nachstehenden Code in Ihre Authenticator-App eingeben.",
         "label": "Secret"
       },
       "title": "2-Faktor Authentifizierung",
@@ -55,9 +55,9 @@ export default {
       "title": "Newsletter",
       "unsubscribe": "Vom Newsletter abmelden"
     },
-    "phoneInfo": "Mit der Angabe deiner Telefonnummer erklärst du dich damit einverstanden, telefonisch kontaktiert zu werden.",
-    "phoneInfoAdmin": "Du musst die Zustimmung der Person einholen, bevor du ihre Telefonnummer in das System eingibst.",
-    "savedPassword": "Das neue Passwort wurde gespeichert. ",
+    "phoneInfo": "Mit der Angabe Ihrer Telefonnummer erklären Sie sich damit einverstanden, telefonisch kontaktiert zu werden.",
+    "phoneInfoAdmin": "Sie müssen die Zustimmung der Person einholen, bevor Sie ihre Telefonnummer in das System eingeben.",
+    "savedPassword": "Das neue Passwort wurde gespeichert.",
     "subTitle": "Alle benötigten Felder sind mit * markiert. Alle Informationen werden vertraulich behandelt.",
     "title": "Account"
   },
@@ -128,7 +128,7 @@ export default {
     "title": "Hinweis hinzufügen"
   },
   "adminDashboard": {
-    "hint1": "<p>Das beabee-Handbuch für Community-Journalismus <a href=\"https://beabee.io/neu-handbuch-community-journalismus/\" target=\"_blank\" rel=\"noopener noreferrer\">„Lust auf Lokal” ist erschienen</a>. Es enthält nicht nur unser gesammeltes Wissen zum Aufbau unabhängiger Lokalredaktionen, sondern auch viele Beispiele aus der Praxis.</p>\n",
+    "hint1": "<p>Das beabee-Handbuch für Community-Journalismus „Lust auf Lokal” <a href=\"https://beabee.io/neu-handbuch-community-journalismus/\" target=\"_blank\" rel=\"noopener noreferrer\">ist erschienen</a>. Es enthält nicht nur unser gesammeltes Wissen zum Aufbau unabhängiger Lokalredaktionen, sondern auch viele Beispiele aus der Praxis.</p>\n",
     "hint2": "<p>Im <a href=\"https://beabee.io/#blog\" target=\"_blank\" rel=\"noopener noreferrer\">Blog</a> auf unserer Website findest du viele Best-Practice-Beispiele anderer Redaktionen, von denen du dich inspierieren lassen kannst.</p>\n",
     "hint3": "<p>Unsere Community-Meetups finden jeden dritten Donnerstag im Monat statt. Du willst dein Best-Practice-Beispiel mit anderen teilen? Schreib eine E-Mail an <a href=\"mailto:tobias.hauswurz{'@'}correctiv.org\">tobias.hauswurz{'@'}correctiv.org</a></p>\n",
     "joinSlack": "Tritt der beabee-Community auf Slack bei!",
@@ -146,15 +146,15 @@ export default {
       "revenue": "Einnahmen",
       "title": "Überblick der letzten 30 Tage"
     },
-    "responsesSoFar": "{n} Antworten bisher",
+    "responsesSoFar": "{n} Antwort bisher | {n} Antworten bisher",
     "seeAllResponses": "Alle Antworten anzeigen",
-    "supportInbox": "<p>Benötigst du Hilfe? Besuche unser <a href=\"https://wiki.beabee.io/beabee-help-center/\" target=\"_blank\" rel=\"noopener noreferrer\">beabee Help Center</a> oder schreib' uns eine E-Mail an:  <a href=\"mailto:support{'@'}beabee.io\">support{'@'}beabee.io</a></p>\n",
+    "supportInbox": "<p>Benötigst du Hilfe? Besuche unser <a href=\"https://wiki.beabee.io/beabee-help-center/\" target=\"_blank\" rel=\"noopener noreferrer\">beabee Help Center</a> oder schreib' uns eine E-Mail an: <a href=\"mailto:support{'@'}beabee.io\">support{'@'}beabee.io</a></p>\n",
     "welcomeBack": "Willkommen zurück {firstName}!"
   },
   "adminSettings": {
     "apikey": {
       "confirmCreate": {
-        "text": "Bitte speichere den folgenden Token. Er kann später NICHT wieder angezeigt werden.",
+        "text": "Bitte speichern Sie den folgenden Token. Er kann später NICHT wieder angezeigt werden.",
         "title": "API Key erfolgreich erstellt"
       },
       "confirmDelete": {
@@ -333,15 +333,15 @@ export default {
     },
     "addAPoint": "Klicken Sie dort auf die Karte, wo Sie einen Ort ergänzen möchten",
     "addLocation": "Neuen Eintrag eingeben",
-    "confirmHuman": "Bitte bestätige, dass du ein Mensch bist",
+    "confirmHuman": "Bitte bestätigen Sie, dass Sie ein Mensch sind",
     "contactDetails": "Kontaktdaten",
     "ended": "Dieser Callout ist seit dem {date} beendet.",
     "entryOf": "Eintrag {no} von {total}",
     "form": {
-      "email": "Deine E-Mail-Adresse",
-      "guestFieldsMissing": "Bitte gib' uns Deine Kontaktdaten. ",
-      "name": "Dein Name",
-      "submittingResponseError": "Es gab ein Problem bei der Übermittlung deiner Antworten. Bitte versuche es erneut."
+      "email": "Ihre E-Mail-Adresse",
+      "guestFieldsMissing": "Bitte geben Sie uns Ihre Kontaktdaten.",
+      "name": "Ihr Name",
+      "submittingResponseError": "Es gab ein Problem bei der Übermittlung ihrer Antworten. Bitte versuchen Sie es erneut."
     },
     "getStarted": "Starten",
     "joinNow": "Mitglied werden",
@@ -354,13 +354,13 @@ export default {
     },
     "showingPreview": "Im Vorschaumodus sind Antworten deaktiviert.",
     "toContributionPage": "Meinen Beitrag ändern",
-    "updateContribution": "Bitte ändere deinen Beitrag. ",
+    "updateContribution": "Bitte ändern Sie ihren Beitrag.",
     "views": {
       "gallery": "Galerie",
       "map": "Karte"
     },
-    "youResponded": "Du hast geantwortet ♥",
-    "yourResponse": "Deine Antwort"
+    "youResponded": "Sie haben geantwortet ♥",
+    "yourResponse": "Ihre Antwort"
   },
   "calloutAdmin": {
     "deleted": "Callout gelöscht",
@@ -432,7 +432,7 @@ export default {
   },
   "calloutResponse": {
     "data": {
-      "answers": "Antworten",
+      "answers": "Anworten",
       "assignee": "Zugewiesen an",
       "bucket": "Ordner",
       "callout": "Callout",
@@ -529,7 +529,7 @@ export default {
   },
   "common": {
     "backToHome": "Zurück zur Startseite",
-    "contributing": "Dein Beitrag",
+    "contributing": "Ihr Beitrag",
     "contributionPeriod": {
       "annually": "Jährlich",
       "monthly": "Monatlich"
@@ -631,7 +631,7 @@ export default {
       "instructions": "Link kopieren und teilen:",
       "loginOverride": "Automatischer Login",
       "mfa": {
-        "desc": "Dieser Kontakt hat 2FA eingeschaltet. Klick auf \"{ disableLabel }\", um es auszuschalten.",
+        "desc": "Dieser Kontakt hat 2FA aktiviert. Klicken Sie auf \"{ disableLabel }\", um es zu deaktivieren.",
         "disabledDesc": "Dieser Kontakt hat keine 2FA aktiviert."
       },
       "resetPassword": "Passwort zurücksetzen",
@@ -663,11 +663,11 @@ export default {
       "joined": "Kontakt seit",
       "lastSeen": "Zuletzt angemeldet",
       "lastname": "Nachname",
-      "manualPaymentSource": "Manuelles Zahlungsmittel",
+      "manualPaymentSource": "Manuelle Zahlungsquelle",
       "membershipExpires": "Mitglied bis",
       "membershipStarts": "Mitglied seit",
       "name": "Name",
-      "newsletterGroups": "Newsletter-Gruppe",
+      "newsletterGroups": "Newsletter groups",
       "newsletterStatus": "Newsletter-Status",
       "notes": "Notizen",
       "payingFee": "Übernimmt Transaktionskosten?",
@@ -696,44 +696,44 @@ export default {
     "bankAccount": "Bankverbindung",
     "billing": "Zahlungsart",
     "cancelContribution": "Mitgliedschaft kündigen",
-    "cancelDescription": "Wenn du jetzt kündigst, endet deine Mitgliedschaft am {date}.",
-    "cancelMessage": "Schade, dass du uns verlässt.",
-    "cancelledContribution": "Deine Mitgliedschaft wurde gekündigt.",
-    "changeBank": "Ändere deine Bankverbindung.",
-    "changeBankInfo": "Im nächsten Schritt kannst du deine neue Bankverbindung eingeben.",
+    "cancelDescription": "Wenn Sie jetzt kündigen, endet Ihre Mitgliedschaft am {date}.",
+    "cancelMessage": "Schade, dass Sie uns verlassen.",
+    "cancelledContribution": "Ihre Mitgliedschaft wurde gekündigt.",
+    "changeBank": "Ändern Sie Ihre Bankverbindung.",
+    "changeBankInfo": "Im nächsten Schritt können Sie Ihre neue Bankverbindung eingeben.",
     "contributing": "Sie zahlen",
-    "contributionUpdateError": "Du kannst deinen Beitrag aktuell nicht ändern, weil eine Zahlung kurz bevorsteht.",
-    "dueToRenew": "Deine Mitgliedschaft wird am {renewalDate} erneuert.",
-    "expired": "Deine Mitgliedschaft ist ausgelaufen.",
-    "expiredText": "Möchtest Du deine Mitgliedschaft verlängern? Selbst kleinste Beträge helfen uns und unserer Arbeit. Danke!",
-    "hasPendingPayment": "Deine Zahlung wird gerade verarbeitet.",
-    "manualPayment": "Du zahlst deinen Beitrag aktuell via {source}. Indem du über unser Mitgliedersystem bezahlst, bekommen wir einen besseren Überblick, auf welche Beiträge wir uns monatlich verlassen können. Für dich wird es außerdem einfacher, deinen Beitrag jederzeit zu ändern.",
+    "contributionUpdateError": "Sie können Ihren Beitrag aktuell nicht ändern, weil eine Zahlung kurz bevorsteht.",
+    "dueToRenew": "Ihre Mitgliedschaft wird am {renewalDate} erneuert.",
+    "expired": "Ihre Mitgliedschaft ist ausgelaufen.",
+    "expiredText": "Möchten Sie Ihre Mitgliedschaft verlängern? Selbst kleinste Beträge helfen uns und unserer Arbeit. Danke!",
+    "hasPendingPayment": "Ihre Zahlung wird gerade verarbeitet.",
+    "manualPayment": "Sie zahlen Ihren Beitrag aktuell via {source}. Indem Sie über unser Mitgliedersystem bezahlen, bekommen wir einen besseren Überblick, auf welche Beiträge wir uns monatlich verlassen können. Für Sie wird es außerdem einfacher, Ihren Beitrag jederzeit zu ändern.",
     "manualPaymentSource": "einem Drittanbieter",
-    "nextAmountChanging": "Ab dem {renewalDate} zahlst Du {nextAmount}.",
+    "nextAmountChanging": "Ab dem {renewalDate} zahlen Sie {nextAmount}.",
     "paymentHistory": {
       "empty": "Ältere Zahlungen und Rechnungen finden Sie hier.",
       "title": "Bisher gezahlte Beiträge"
     },
     "paymentMethod": "Zahlungsart",
-    "paymentSourceUpdateError": "Du kannst deine Bankverbindung aktuell nicht ändern, weil eine Zahlung kurz bevorsteht.",
-    "prorateDecreaseMessage": "Dein Beitrag bleibt noch für 1 Monat gleich. | Dein Beitrag bleibt noch für {n} Monaten gleich.",
+    "paymentSourceUpdateError": "Sie können Ihre Bankverbindung aktuell nicht ändern, weil eine Zahlung kurz bevorsteht.",
+    "prorateDecreaseMessage": "Ihr Beitrag bleibt noch für 1 Monat gleich. | Ihr Beitrag bleibt noch für {n} Monaten gleich.",
     "prorateDecreaseMessageWillStart": "Der neue Beitrag wird am {renewalDate} zum ersten Mal eingezogen.",
-    "prorateIncreaseMessage": "Deine Mitgliedschaft mit dem oben angezeigten Beitrag läuft noch 1 Monat. | Deine Mitgliedschaft mit dem oben angezeigten Beitrag läuft noch {n} Monaten.",
-    "prorateIncreaseMessageOneOff": "Du kannst aber eine Einmalzahlung in Höhe von {oneOffPayment} leisten um deine Mitgliedschaft sofort zu verlängern und den Beitrag zu erhöhen.",
-    "prorateNoText": "Wählst Du diese Option, bleibt dein Beitrag bis zum neuen Abrechnungszeitrarum gleich.",
+    "prorateIncreaseMessage": "Ihre Mitgliedschaft mit dem oben angezeigten Beitrag läuft noch 1 Monat. | Ihre Mitgliedschaft mit dem oben angezeigten Beitrag läuft noch {n} Monaten.",
+    "prorateIncreaseMessageOneOff": "Sie können aber eine Einmalzahlung in Höhe von {oneOffPayment} leisten um Ihre Mitgliedschaft sofort zu verlängern und den Beitrag zu erhöhen.",
+    "prorateNoText": "Wählen Sie diese Option, bleibt Ihr Beitrag bis zum neuen Abrechnungszeitrarum gleich.",
     "prorateNoTitle": "Ändern Sie meinen Beitrag ab dem {renewalDate}.",
-    "prorateYesText": "Wählst Du diese Option, buchen wir eine Einmalzahlung in Höhe von {oneOffPayment} ab und deine Mitgliedschaft wird mit deinem neuen Beitrag sofort verlängert.",
+    "prorateYesText": "Wählen Sie diese Option, buchen wir eine Einmalzahlung in Höhe von {oneOffPayment} ab und Ihre Mitgliedschaft wird mit Ihrem neuen Beitrag sofort verlängert.",
     "prorateYesTitle": "Meinen Beitrag jetzt erhöhen",
     "restartContribution": "Mitgliedschaft erneuern",
     "startContribution": "Mitgliedschaft beginnen",
-    "startedContribution": "Deine Mitgliedschaft hat begonnen.",
+    "startedContribution": "Ihre Mitgliedschaft hat begonnen.",
     "updateContribution": "Beitrag aktualisieren",
     "updatePaymentType": "Neue Zahlungsmethode einrichten",
-    "updatedContribution": "Dein Beitrag wurde geändert.",
-    "updatedPaymentSource": "Deine Zahlungsart wurde aktualisiert. ",
-    "willExpire": "Deine Mitgliedschaft läuft in {expires} ab.",
-    "willRenew": "Deine Mitgliedschaft verlängert sich am {renewalDate} automatisch.",
-    "willRenewWithNewAmount": "Deine Mitgliedschaft verlängert sich am {renewalDate} automatisch. Dein Beitrag ist dann {nextAmount}."
+    "updatedContribution": "Ihr Beitrag wurde geändert.",
+    "updatedPaymentSource": "Ihre Zahlungsart wurde aktualisiert.",
+    "willExpire": "Ihre Mitgliedschaft läuft in {expires} ab.",
+    "willRenew": "Ihre Mitgliedschaft verlängert sich am {renewalDate} automatisch.",
+    "willRenewWithNewAmount": "Ihre Mitgliedschaft verlängert sich am {renewalDate} automatisch. Ihr Beitrag ist dann {nextAmount}."
   },
   "createCallout": {
     "lastSaved": "Zuletzt vor {duration} gespeichert.",
@@ -776,7 +776,7 @@ export default {
             "placeholder": "Erkläre den Teilnehmenden, warum ihre Antwort hilfreich ist und was damit passiert."
           },
           "title": {
-            "help": "<p>Die Teilnehmer sehen den <strong>Titel</strong> wenn sie ihre Antworten abgeschickt haben.</p>\n",
+            "help": "<p>Die Teilnehmenden sehen den <strong>Titel der Dankesnachricht</strong> wenn sie ihre Antworten abgeschickt haben.</p>\n",
             "label": "Titel der Dankesnachricht",
             "placeholder": "Danke!"
           },
@@ -970,7 +970,7 @@ export default {
     "title": "Bearbeiten: {title}"
   },
   "footer": {
-    "contactUs": "Kontaktiere uns unter",
+    "contactUs": "Kontaktieren Sie uns unter",
     "handbookResources": "Handbuch & Ressourcen",
     "impressum": "Impressum",
     "joinCommunity": "Mitglied werden",
@@ -980,8 +980,8 @@ export default {
   },
   "forgotPassword": {
     "backToLogin": "Zurück zum Login",
-    "description": "Bitte gib deine E-Mail-Adresse ein. Wir schicken dir dann die Anleitung zum Zurücksetzen deines Passworts.",
-    "message": "Gibt es einen Account mit der E-mail-Adresse {email}, bekommst du in Kürze eine Nachricht. Bitte schaue in deinen Posteingang und befolge die Anleitung in der E-Mail. Sollte die E-mail nicht auftauchen, checke deinen Spamordner. ",
+    "description": "Bitte geben Sie Ihre E-Mail-Adresse ein. Wir schicken Ihnen die Anleitung zum Zurücksetzten Ihres Passworts.",
+    "message": "Gibt es einen Account mit der E-Mail-Adresse {email}, bekommen Sie in Kürze eine Nachricht. Bitte schauen Sie in Ihren Posteingang und befolgen Sie die Anleitung in der E-Mail. Sollte die E-Mail nicht auftauchen, schauen Sie auch in ihrem Spamordner nach.",
     "title": "Passwort vergessen?"
   },
   "form": {
@@ -997,17 +997,17 @@ export default {
         "account-locked": "Konto wegen zu vieler fehlgeschlagener Anmeldeversuche gesperrt. Bitte Passwort zurücksetzen oder an den Support wenden, um Hilfe zu erhalten.",
         "duplicate-email": "Diese E-Mail-Adresse existiert bereits.",
         "duplicate-tag-name": "Ein Tag mit diesem Namen existiert bereits",
-        "invalid-token": "Dein eingegebener 2FA Token stimmt nicht.",
+        "invalid-token": "Der von Ihnen eingegebene 2FA Token ist nicht korrekt.",
         "login-failed": "Falsche Kombination von Passwort und E-Mail.",
-        "mfa-token-required": "Bitte gebe den Code deiner Authentifizierungs-App ein."
+        "mfa-token-required": "Bitte geben Sie den Code Ihrer Authentifizierungs-App ein."
       },
-      "generic": "Upps! Etwas ist schiefgelaufen. Bitte wende dich an den Support, sollte das erneut passieren.",
-      "unauthorized": "Du hast daruf keinen Zugriff.",
-      "validation": "Es fehlt etwas! Bitte kontrolliere die obigen Felder."
+      "generic": "Upps! Etwas ist schiefgelaufen. Bitte wenden Sie sich an den Support, sollte das erneut passieren.",
+      "unauthorized": "Sie haben daruf keinen Zugriff.",
+      "validation": "Es fehlt etwas! Bitte kontrollieren Sie die obigen Felder."
     },
     "errors": {
       "address": {
-        "required": "Dieses Feld wird benötigt, wenn du eine Adresse eintragen willst."
+        "required": "Dieses Feld wird benötigt, wenn Sie eine Adresse eintragen wollen."
       },
       "addressLine1": {
         "required": "Adresse benötigt"
@@ -1029,7 +1029,7 @@ export default {
         "required": "E-Mail-Adresse benötigt"
       },
       "file": {
-        "rateLimited": "Hochladen fehlgeschlagen: Du hast in zu kurzer Zeit zu viele Dateien hochgeladen. Bitte warte ein paar Minuten, bevor du es erneut versuchst.",
+        "rateLimited": "Hochladen fehlgeschlagen: Sie haben in zu kurzer Zeit zu viele Dateien hochgeladen. Bitte warten Sie ein paar Minuten, bevor Sie es erneut versuchen.",
         "tooBig": "Hochladen fehlgeschlagen: Die Dateigröße darf max. 20 MB betragen."
       },
       "firstName": {
@@ -1061,8 +1061,8 @@ export default {
         "url": "Ungültige URL"
       },
       "verifyCode": {
-        "max": "Dein Code darf höchstens {max} Zeichen lang sein.",
-        "min": "Dein Code muss mindestens {min} Zeichen haben.",
+        "max": "Der Code darf maximal {max} Zeichen lang sein.",
+        "min": "Der Code muss mindestens {min} Zeichen lang sein.",
         "required": "Code benötigt"
       }
     },
@@ -1114,8 +1114,8 @@ export default {
     "title": "Home",
     "viewAllCallouts": "Alle Callouts anzeigen",
     "viewOpenCallouts": "Aktive Callouts anzeigen",
-    "welcome": "Herzlich willkommen {firstName}!",
-    "yourProfile": "Dein Profil"
+    "welcome": "Herzlich willkommen {firstName} {lastName}!",
+    "yourProfile": "Ihr Profil"
   },
   "item": {
     "status": {
@@ -1124,28 +1124,28 @@ export default {
     }
   },
   "join": {
-    "absorbFeeForce": "Bei einem Beitrag von {amount}, musst du die {fee} Transaktionskosten übernehmen.",
+    "absorbFeeForce": "Bei einem Beitrag von {amount}, müssen Sie die {fee} Transaktionskosten übernehmen.",
     "absorbFeeOptIn": "Ja, ich übernehme die {fee} Transaktionskosten.",
-    "absorbFeeText": "Unser Zahlungsdienstleister stellt uns jede Transaktion in Rechnung. Das heißt, wir erhalten weniger von deinem monatlichen Beitrag. Möchtest du die Kosten dafür übernehmen? Alternativ kannst du auch jährlich bezahlen.",
+    "absorbFeeText": "Bei unserem Zahlungsdienstleister zahlen wir für jede einzelne Transaktion. Das heißt, wir erhalten weniger von Ihrem monatlichen Beitrag. Sie haben die Möglichkeit, die Kosten dafür für uns zu übernehmen. Möchten Sie das? Alternativ können Sie auch jährlich bezahlen.",
     "contribute": "Mit {contribution} unterstützen.",
     "errors": {
       "confirmEmail": {
-        "message": "Fast geschafft! Wir müssen nur noch deine E-Mail-Adresse bestätigen, um fortzufahren.",
-        "proceed": "Wir haben dir eine E-Mail geschickt. Bitte klicke auf den Bestätigungslink, um fortzufahren.",
+        "message": "Fast geschafft! Wir müssen nur noch Ihre E-Mail-Adresse bestätigen, um fortzufahren.",
+        "proceed": "Wir haben Ihnen eine E-Mail geschickt. Bitte klicken Sie auf den Bestätigung-Link, um fortzufahren.",
         "title": "E-Mail-Adresse bestätigen"
       },
       "failed": {
         "backToJoin": "Zurück zur Registrierung",
-        "message": "Es gab ein Problem bei der Registrierung, unser Team wurde benachrichtigt. Bitte kontaktiere uns, wenn das Problem weiterhin auftritt.",
+        "message": "Es gab ein Problem bei der Registrierung, unser Team wurde benachrichtigt. Bitte kontaktieren Sie uns, wenn das Problem weiterhin auftritt.",
         "title": "Etwas ist schief gelaufen"
       }
     },
     "login": "Anmelden",
     "memberAccount": "Mitglieder-Account",
-    "memberAlready": "Bist du bereits Mitglied?",
+    "memberAlready": "Sind Sie bereits Mitglied?",
     "minimumContribution": "Der minimale Beitrag ist",
     "noContribution": "Ich möchte jetzt keinen Beitrag leisten",
-    "notice": "Indem du fortfährst, akzeptierst Du unsere",
+    "notice": "Indem Sie fortfahren, akzeptieren Sie unsere",
     "now": "Jetzt Mitglied werden! ",
     "paymentMethod": "Zahlungsart",
     "poweredBy": "Powered by beabee",
@@ -1159,16 +1159,16 @@ export default {
       "description": "Mitgliedschaft",
       "recurringLabel": "Regelmäßiger Beitrag"
     },
-    "genericError": "Upps! Etwas ist schiefgelaufen. Bitte wende dich an den Support, sollte das erneut passieren.",
-    "goBack": "Möchtest du etwas ändern? {back}",
+    "genericError": "Upps! Etwas ist schiefgelaufen. Bitte wenden Sie sich an den Support, sollte das erneut passieren.",
+    "goBack": "Möchten Sie etwas ändern? {back}",
     "goBackButton": "Zurück zum letzten Schritt",
-    "note": "Du kannst deine Zahlungsdaten jederzeit in deinem Nutzer-Account ändern und dort auch sehen, wie viel du in der Vergangenheit bereits gezahlt hast. ",
-    "willBeContributing": "Du zahlst {contribution}"
+    "note": "Sie können Ihre Zahlungsdaten jederzeit in Ihrem Nutzer-Account ändern und dort auch sehen, wie viel Sie in der Vergangenheit bereits gezahlt haben. ",
+    "willBeContributing": "Sie zahlen {contribution}"
   },
   "joinSetup": {
-    "confirmDetails": "Bitte bestätige deine Angaben:",
+    "confirmDetails": "Bitte bestätigen Sie Ihre Angaben:",
     "continue": "Weiter",
-    "welcome": "Willkommen {firstName}!"
+    "welcome": "Willkommen {firstName} {lastName}!"
   },
   "login": {
     "backTo": "Zurück zu {newsroomName}",
@@ -1179,9 +1179,9 @@ export default {
     "title": "Einloggen"
   },
   "lostDevice": {
-    "description": "Bitte gib deine E-Mail-Adresse ein, und wir senden dir Anweisungen, wie du dein 2FA zurücksetzen kannst.",
-    "message": "Gibt es einen Account mit der E-mail-Adresse {email}, bekommst du in Kürze eine Nachricht. Bitte schaue in deinen Posteingang und befolge die Anleitung in der E-Mail. Sollte die E-mail nicht auftauchen, checke deinen Spamordner. ",
-    "title": "Du hast dein 2FA-Gerät verloren?"
+    "description": "Bitte geben Sie Ihre E-Mail-Adresse ein, und wir senden Ihnen Anweisungen zum Zurücksetzen Ihres 2FA.",
+    "message": "Gibt es einen Account mit der E-Mail-Adresse {email}, bekommen Sie in Kürze eine Nachricht. Bitte schauen Sie in Ihren Posteingang und befolgen Sie die Anleitung in der E-Mail. Sollte die E-Mail nicht auftauchen, schauen Sie auch in ihrem Spamordner nach.",
+    "title": "Haben Sie Ihr 2FA-Gerät verloren?"
   },
   "membershipBuilder": {
     "steps": {
@@ -1195,7 +1195,7 @@ export default {
           "required": "Die Mitglieder müssen den Fragebogen ausfüllen, ehe sie fortfahren können.",
           "slug": "Callout",
           "slugHelp": "Du kannst jeden Callout verwenden, der Enddatum hat.",
-          "text": "Als Teil des Anmeldeprozesses kannst du deine Mitglieder einen kurzen Fragebogen ausfüllen lassen. Du kannst zum Beispiel zusätzliche Informationen abfragen oder in Erfahrung bringen, warum sie Mitglied geworden sind.",
+          "text": "Als Teil des Anmeldeprozesses kannst du deine neuen Mitglieder einen kurzen Fragebogen ausfüllen lassen. Du kannst zum Beispiel zusätzliche Informationen abfragen oder in Erfahrung bringen, warum sie Mitglied geworden sind.",
           "textIntro": "Erkläre deinen Mitgliedern kurz, warum du weitere Informationen von ihnen haben möchtest.",
           "title": "Anmelde-Fragebogen"
         },
@@ -1262,7 +1262,7 @@ export default {
     "notices": "Hinweise",
     "payments": "Zahlungen",
     "settings": "Einstellungen",
-    "yourProfile": "Dein Profil"
+    "yourProfile": "Ihr Profil"
   },
   "notFound": {
     "goBack": "Zurück zum Dashboard",
@@ -1298,7 +1298,7 @@ export default {
     "forgotPassword": "Passwort vergessen",
     "join": "Mitglied werden",
     "login": "Login",
-    "lostDevice": "Du hast dein 2FA-Gerät verloren?",
+    "lostDevice": "Haben Sie Ihr 2FA-Gerät verloren?",
     "notFound": "Seite nicht gefunden",
     "resetDevice": "Zweitgerät zurücksetzen",
     "resetPassword": "Passwort zurücksetzen",
@@ -1309,32 +1309,32 @@ export default {
     "gc_direct-debit": {
       "changeLabel": "Lastschrift ändern",
       "label": "Lastschrift",
-      "setLabel": "Deine Laschriftdaten"
+      "setLabel": "Ihre Laschriftdaten"
     },
     "s_bacs": {
       "changeLabel": "Lastschrift ändern",
       "label": "Lastschrift",
-      "setLabel": "Deine Laschriftdaten"
+      "setLabel": "Ihre Laschriftdaten"
     },
     "s_card": {
       "changeLabel": "Kreditkartendetails ändern",
       "label": "Kreditkarte",
-      "setLabel": "Deine Kreditkartendetails"
+      "setLabel": "Ihre Kreditkartendetails"
     },
     "s_ideal": {
       "changeLabel": "iDEAL ändern",
       "label": "iDEAL",
-      "setLabel": "Deine iDEAL-Daten"
+      "setLabel": "Ihre iDEAL-Daten"
     },
     "s_paypal": {
       "changeLabel": "PayPal ändern",
       "label": "PayPal",
-      "setLabel": "Deine PayPal-Daten"
+      "setLabel": "Ihre PayPal-Daten"
     },
     "s_sepa": {
       "changeLabel": "Lastschrift ändern",
       "label": "Lastschrift",
-      "setLabel": "Deine Laschriftdaten"
+      "setLabel": "Ihre Laschriftdaten"
     }
   },
   "payments": {
@@ -1359,18 +1359,18 @@ export default {
     "showingOf": "Zahlungen {start} bis {end} von {total}"
   },
   "resetDevice": {
-    "description": "Gib deinPasswort ein, um den 2FA-Schutz zurückzusetzen. Sobald du eingeloggt bist, kannst du ein neues Gerät einrichten.",
+    "description": "Geben Sie Ihr Passwort ein, um den 2FA-Schutz zurückzusetzen. Sobald Sie eingeloggt sind, können Sie ein neues Gerät einrichten.",
     "failed": "Das Zurücksetzen des Geräts war nicht erfolgreich, möglicherweise ist die Verbindung abgelaufen.",
     "invalidPassword": "Ungültiges Passwort",
-    "success": "Das Zurücksetzen des 2FA-Geräts hat geklappt.",
+    "success": "Das Zurücksetzen des 2FA-Gerätes war erfolgreich.",
     "title": "2FA-Gerät zurücksetzen?"
   },
   "resetPassword": {
     "description": "Neues Passwort eingeben, um fortzufahren.",
-    "failed": "Das Zurücksetzen des Passworts ist schiefgelaufen.",
+    "failed": "Das Zurücksetzen des Passworts ist schiefgelaufen",
     "login": "Einloggen",
     "lostDevice": "2FA-Gerät verloren? Bitte an den Support wenden.",
-    "success": "Dein Passwort wurde aktualisiert.",
+    "success": "Ihr Passwort wurde aktualisiert.",
     "title": "Passwort aktualisieren?"
   },
   "roleEditor": {
@@ -1404,9 +1404,9 @@ export default {
   "tagEditor": {
     "add": "Tag hinzufügen",
     "confirmDelete": {
-      "text_contact": "Bist du sicher, dass du diesen Tag entfernen willst? Jeder Kontakt mit dem Tag \"{tagName}\" wird den Tag verlieren.",
-      "text_global": "Bist du sicher, dass du den Tag \"{tagName}\" entfernen willst?",
-      "text_response": "Bist du sicher, dass du diesen entfernen willst? Jede Antwort mit dem Tag \"{tagName}\" wird den Tag verlieren.",
+      "text_contact": "Sind Sie sicher, dass Sie diesen Tag entfernen wollen? Jeder Kontakt mit dem Tag \"{tagName}\" wird den Tag verlieren.",
+      "text_global": "Sind Sie sicher, dass Sie den Tag \"{tagName}\" entfernen wollen?",
+      "text_response": "Sind Sie sicher, dass Sie diesen Tag entfernen wollen? Jede Antwort mit dem Tag \"{tagName}\" wird den Tag verlieren.",
       "title": "Tag entfernen"
     },
     "name": "Name"
