+++ packages/locale/src/locales/de@easy.ts
@@ -1,5 +1,5 @@
 // This file is generated by generate.js
-export default {
+export const deEasy = {
   "accountPage": {
     "mfa": {
       "codeInput": {},
