+++ packages/locale/src/locales/en.ts
@@ -1,5 +1,5 @@
 // This file is generated by generate.js
-export default {
+export const en = {
   "accountPage": {
     "contactInformation": "Contact information",
     "deliveryAddress": "Delivery address",
