+++ packages/locale/src/locales/index.ts
@@ -1,0 +1,9 @@
+// This file is generated by generate.js
+export * from './en.js';
+export * from './de.js';
+export * from './de@informal.js';
+export * from './pt.js';
+export * from './it.js';
+export * from './ru.js';
+export * from './nl.js';
+export * from './de@easy.js';
