+++ packages/locale/src/locales/it.ts
@@ -1,5 +1,5 @@
 // This file is generated by generate.js
-export default {
+export const it = {
   "accountPage": {
     "contactInformation": "Informazioni di contatto",
     "deliveryAddress": "Indirizzo",
