+++ packages/locale/src/locales/nl.ts
@@ -1,5 +1,5 @@
 // This file is generated by generate.js
-export default {
+export const nl = {
   "accountPage": {
     "contactInformation": "Contactgegevens",
     "deliveryAddress": "Afleveradres",
