+++ packages/locale/src/locales/pt.ts
@@ -1,5 +1,5 @@
 // This file is generated by generate.js
-export default {
+export const pt = {
   "accountPage": {
     "contactInformation": "Informação de contacto",
     "deliveryAddress": "Endereço de entrega",
