+++ packages/locale/src/locales/ru.ts
@@ -1,5 +1,5 @@
 // This file is generated by generate.js
-export default {
+export const ru = {
   "accountPage": {
     "contactInformation": "Контакты",
     "deliveryAddress": "Адрес",
